Add files via upload
Task 6: For the person object, add a method named greet that uses the this keyword to greet with the person’s name. 
Task 7: Design a circle object with properties radius and a method to calculate its area using this.radius. 
Task 8: Extend the circle object with methods to calculate its diameter and circumference. 
Task 9: Create an object account with properties: name, balance and methods: deposit, withdraw. Use the this keyword appropriately. 
Task 10: For the account object, ensure that the balance can’t go negative using the this keyword.